requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
pinecone-client==5.0.1
sentence-transformers==3.1.1
openai==1.46.0
tiktoken==0.7.0
langdetect==1.0.9
pydantic==2.9.2